+++ CHANGELOG.md
@@ -7,8 +7,9 @@ Tidarr notable changes.
 ### 🚀 Added
 - [Front] Add API key manager in configuration dialog
 - [API] Add X-API-KEY header to authenticate external queries
-- [API] Add Lidarr NZB indexer endpoints
 - [API] Add replay gain post processing step (REPLAY_GAIN=true)
+- [API] BETA - Add Lidarr NZB indexer endpoints
+- [API] BETA - Add Lidarr Sabnzbd download client endpoints
 
 ## 📦 1.1.0
 ### 🚀 Added